Add tags for programs
