+++ src/pages/ProgramsList/ProgramItems.js
@@ -2,6 +2,7 @@ import React, { Fragment, useState } from "react";
 import { Input, Layout, List, Tabs, PageHeader } from "antd";
 import ItemDetailsModal from "./ItemDetailsModal";
 import ListItem from "./ListItem";
+import ProgramsItemsSider from "./ProgramItemsSider";
 
 const { Content } = Layout;
 const { TabPane } = Tabs;
@@ -18,6 +19,7 @@ for (let i = 0; i < 23; i++) {
       venue: "NTU North Spine ...",
       duration: "2 hours",
     },
+    tags: ["Food", "Lifestyle"],
     avatar: "https://image.flaticon.com/icons/svg/3163/3163231.svg",
     description:
       "This is part of a research study to investigate the correlation between eating habits and stress levels.",
@@ -43,55 +45,58 @@ const ProgramItems = () => {
 
   return (
     <Fragment>
-      <PageHeader
-        title="Programs"
-        className="site-page-header"
-        extra={
-          <Tabs
-            defaultActiveKey="1"
-            selectedKeys={current}
-            onClick={handleClick}
-          >
-            <TabPane tab="All Programs" key="1" />
-            <TabPane tab="Registered Programs" key="2" />
-          </Tabs>
-        }
-      >
-        <Search
-          placeholder="input search text"
-          onSearch={(value) => console.log(value)}
-          style={{ width: 700 }}
-        />
-        <Content
-          className="site-layout-background"
-          style={{
-            padding: 24,
-            margin: 0,
-          }}
+      <Layout style={{ marginLeft: 200 }}>
+        <ProgramsItemsSider />
+        <PageHeader
+          title="Programs"
+          className="site-page-header"
+          extra={
+            <Tabs
+              defaultActiveKey="1"
+              selectedKeys={current}
+              onClick={handleClick}
+            >
+              <TabPane tab="All Programs" key="1" />
+              <TabPane tab="Registered Programs" key="2" />
+            </Tabs>
+          }
         >
-          <List
-            itemLayout="vertical"
-            size="large"
-            pagination={{
-              onChange: (page) => {
-                console.log(page);
-              },
-              pageSize: 7,
-            }}
-            dataSource={dummyData}
-            renderItem={(item) => (
-              <ListItem item={item} showModal={showModal} />
-            )}
+          <Search
+            placeholder="input search text"
+            onSearch={(value) => console.log(value)}
+            style={{ width: 700 }}
           />
-          {isModalVisible && (
-            <ItemDetailsModal
-              isModalVisible={isModalVisible}
-              setModalVisible={setModalVisible}
-              item={modalItem}
+          <Content
+            className="site-layout-background"
+            style={{
+              padding: 24,
+              margin: 0,
+            }}
+          >
+            <List
+              itemLayout="vertical"
+              size="large"
+              pagination={{
+                onChange: (page) => {
+                  console.log(page);
+                },
+                pageSize: 7,
+              }}
+              dataSource={dummyData}
+              renderItem={(item) => (
+                <ListItem item={item} showModal={showModal} />
+              )}
             />
-          )}
-        </Content>
-      </PageHeader>
+            {isModalVisible && (
+              <ItemDetailsModal
+                isModalVisible={isModalVisible}
+                setModalVisible={setModalVisible}
+                item={modalItem}
+              />
+            )}
+          </Content>
+        </PageHeader>
+      </Layout>
     </Fragment>
   );
 };
